Fix definition of DataAuthorization and refactor
